ckan/ckanext/ckanext-ytp_tasks/dev-requirements.txt
pytest-ckan
pytest-cov
mock